+++ odoo/songs/install/data_demo.py
@@ -78,6 +78,12 @@ def import_product(ctx):
 
 
 @anthem.log
+def import_serial_number(ctx):
+    """ Importing serial number from csv """
+    load_csv(ctx, 'data/demo/serial.csv', 'stock.production.lot')
+
+
+@anthem.log
 def import_workcenter(ctx):
     """ Importing workcenters from csv """
     load_csv(ctx, 'data/demo/workcenter.csv', 'mrp.workcenter')
@@ -109,6 +115,7 @@ def main(ctx):
     import_customers(ctx)
     import_drone_type(ctx)
     import_product(ctx)
+    import_serial_number(ctx)
     import_workcenter(ctx)
     import_rma_cause(ctx)
     import_bank(ctx)